Added a listOfStudents, singleStudent and a students component
Added a database name in server.js. Created a listOfStudents component that goes off to the server to pull a new list of students. Created a singleStudent component that displays details on a particular student in the listOFStudents page, including a edit and delete button for each student. The student component takes a collection of students and splits them into individual students