Created an edit component and added form alerts
Added in an edit component for editing indivdual students. Added in an 2 alerts for when the user is signing up for the first time and when the user edits an existing student. These alerts only occur when a field in the form is empty. Also got rid of the id field for student due to connection issues with the database.